feat: Complete Phase 15 - Demo & Portfolio
🎯 Phase 15 Deliverables:
✅ Demo account setup (demo@demo.com / password123)
✅ Seed data generation (5 categories, 30 transactions)
✅ Comprehensive README with setup guide and features
✅ Assets structure for screenshots and GIFs
✅ Production deployment scripts
✅ Portfolio cleanup utilities

📋 Key Features Added:
- Demo data seeding script with realistic transactions
- Enhanced README with demo credentials and features
- Production deployment automation
- Screenshot and GIF asset structure
- Cleanup scripts for unused files
- Complete documentation overhaul

🔧 Technical Improvements:
- Enhanced authentication system
- Improved error handling and logging
- Production-ready configuration
- Comprehensive testing setup
- CI/CD pipeline improvements

🚀 Production Ready:
- Demo account for easy testing
- Complete feature documentation
- Deployment automation
- Asset management structure
- Portfolio presentation ready

This completes all 15 phases of the Flutter Expense Tracker project!

+++ backend/app/models/category.py
@@ -18,7 +18,7 @@ class Category(Base):
     type = Column(Enum(CategoryType), nullable=False)
     icon = Column(String(10), nullable=True)  # Emoji or icon identifier
     color = Column(String(7), nullable=True)  # Hex color code
-    user_id = Column(BigInteger, ForeignKey("users.id"), nullable=False)
+    user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), server_default=func.now(), onupdate=func.now())
